requirements.txt
asgiref==3.6.0
blinker==1.4
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.32.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
Django==4.2.1
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
oauthlib==3.2.0
Pillow==9.5.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0
pytz==2023.3
PyYAML==5.4.1
requests==2.30.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
wadllib==1.3.6
zipp==1.0.0
